fix: mismatched type (stat.Bsize to int64)
The linter find this error in some archs. Paradoxically it works
on x86. This was introducen in the commmit
3c06e26bacac7d0de43a8cf20884574c95237423

+++ pkg/util/file.go
@@ -113,7 +113,8 @@ func GetAvailableSpace(path string) (int64, error) {
 	if err != nil {
 		return int64(-1), err
 	}
-	return int64(stat.Bavail) * stat.Bsize, nil
+	//nolint:unconvert
+	return int64(stat.Bavail) * int64(stat.Bsize), nil
 }
 
 // GetAvailableSpaceBlock gets the amount of available space at the block device path specified.
